Move all helper functions to one file
